+++ projects/netgrif-components-core/src/lib/event/services/event.service.ts
@@ -1,6 +1,6 @@
 import {Injectable} from '@angular/core';
 import {EventOutcome} from '../../resources/interface/event-outcome';
-import {ChangedFields} from '../../data-fields/models/changed-fields';
+import {ChangedFields, FrontAction} from '../../data-fields/models/changed-fields';
 import {SetDataEventOutcome} from '../model/event-outcomes/data-outcomes/set-data-event-outcome';
 import {ChangedFieldsMap} from './interfaces/changed-fields-map';
 import {EventConstants} from '../model/event-constants';
@@ -53,6 +53,30 @@ export class EventService {
             }
         });
         return changedFieldsMap;
+    }
+
+    public parseFrontActionsFromOutcomeTree(outcome: EventOutcome): Array<FrontAction> {
+        const frontActions: Array<FrontAction> = [];
+        if (!!outcome.outcomes && outcome.outcomes.length > 0) {
+            return this.parseFrontActionsFromOutcomeTreeRecursive(outcome.outcomes, frontActions);
+        } else return frontActions;
+    }
+
+    private parseFrontActionsFromOutcomeTreeRecursive(outcomes: Array<EventOutcome>,
+                                                      frontActions: Array<FrontAction>): Array<FrontAction> {
+        outcomes.forEach(childOutcome => {
+            if (EventConstants.FRONT_ACTIONS in childOutcome) {
+                const childFrontActions: Array<FrontAction> = (childOutcome as EventOutcome).frontActions;
+
+                if (!!childFrontActions) {
+                    frontActions.push(...childFrontActions)
+                }
+            }
+            if (!!childOutcome.outcomes && childOutcome.outcomes.length > 0) {
+                this.parseFrontActionsFromOutcomeTreeRecursive(childOutcome.outcomes, frontActions);
+            }
+        });
+        return frontActions;
     }
 
 }