+++ projects/netgrif-components-core/src/lib/public/factories/public-search-factory.ts
@@ -8,14 +8,25 @@ import {FilterType} from '../../filter/models/filter-type';
 import {TranslateService} from '@ngx-translate/core';
 import {PublicTaskLoadingService} from '../../view/task-view/service/public-task-loading.service';
 
+export const CASE_ID = "caseId";
+export const PETRI_NET_ID = "petriNetId"
+export const TRANSITION_ID = "transitionId"
+
 export const publicBaseFilterFactory = (router: Router, route: ActivatedRoute, process: ProcessService,
                                         caseResourceService: CaseResourceService, snackBarService: SnackBarService,
                                         translate: TranslateService, publicTaskLoadingService: PublicTaskLoadingService) => {
-    if (route.snapshot.paramMap.get('caseId') === null && route.snapshot.paramMap.get('petriNetId') !== null) {
+    const caseId = route.snapshot.paramMap.get(CASE_ID)
+    const petriNetId = route.snapshot.paramMap.get(PETRI_NET_ID)
+    const transId = route.snapshot.paramMap.get(TRANSITION_ID)
+    if (caseId === null && petriNetId !== null) {
         getNetAndCreateCase(router, route, process, caseResourceService, snackBarService, translate, publicTaskLoadingService);
-    } else if (route.snapshot.paramMap.get('caseId') !== null) {
+    } else if (caseId !== null && transId !== null) {
         return {
-            filter: new SimpleFilter('', FilterType.TASK, {case: {id: route.snapshot.paramMap.get('caseId')}})
+            filter: new SimpleFilter('', FilterType.TASK, {case: {id: caseId}, transitionId: transId})
+        };
+    } else if (caseId !== null) {
+        return {
+            filter: new SimpleFilter('', FilterType.TASK, {case: {id: caseId}})
         };
     }
     return {