+++ projects/netgrif-components/src/lib/navigation/group-navigation-component-resolver/default-components/default-case-ref-list-view/default-case-ref-list-view.component.ts
@@ -10,6 +10,7 @@ import {
     FilterType,
     NAE_CASE_REF_CREATE_CASE,
     NAE_CASE_REF_SEARCH,
+    CaseRefField,
     NAE_SEARCH_CATEGORIES,
     NAE_TAB_DATA,
     NAE_VIEW_ID_SEGMENT,
@@ -18,7 +19,7 @@ import {
     SearchService,
     SimpleFilter,
     TaskSetDataRequestFields,
-    ViewIdService
+    ViewIdService, DATA_FIELD_PORTAL_DATA, DataFieldPortalData, MultichoiceField, EnumerationField
 } from '@netgrif/components-core';
 import {HeaderComponent} from '../../../../header/header.component'
 import {DefaultTabbedTaskViewComponent} from '../default-tabbed-task-view/default-tabbed-task-view.component';
@@ -60,6 +61,7 @@ export class DefaultCaseRefListViewComponent extends AbstractCaseViewComponent i
     constructor(caseViewService: CaseViewService,
                 @Optional() overflowService: OverflowService,
                 @Optional() @Inject(NAE_TAB_DATA) protected _injectedTabData: InjectedTabbedTaskViewDataWithNavigationItemTaskData,
+                @Optional() @Inject(DATA_FIELD_PORTAL_DATA) protected _dataFieldPortalData: DataFieldPortalData<MultichoiceField | CaseRefField | EnumerationField>,
                 @Optional() @Inject(NAE_CASE_REF_CREATE_CASE) protected _caseRefCreateCase: boolean = false,
                 @Optional() @Inject(NAE_CASE_REF_SEARCH) protected _caseRefSearch: boolean = false) {
         super(caseViewService, overflowService, undefined, {
@@ -74,10 +76,18 @@ export class DefaultCaseRefListViewComponent extends AbstractCaseViewComponent i
         this.initializeHeader(this.caseHeaderComponent);
     }
 
+    public isApproval() {
+        return this._dataFieldPortalData?.dataField instanceof MultichoiceField || this._dataFieldPortalData?.dataField instanceof EnumerationField;
+    }
+
     public handleCaseClick(clickedCase: Case): void {
         if (this._injectedTabData !== null) {
             this.openTab(clickedCase);
         }
+    }
+
+    public disabled(): boolean {
+        return this._dataFieldPortalData?.dataField?.formControlRef.disabled;
     }
 
     protected openTab(openCase: Case) {
@@ -105,5 +115,12 @@ export class DefaultCaseRefListViewComponent extends AbstractCaseViewComponent i
             order: this._injectedTabData['tabViewOrder'],
             parentUniqueId: this._injectedTabData.tabUniqueId
         }, true, true);
+    }
+
+    createdCase(caze: Case) {
+        this.handleCaseClick(caze);
+        // if (this._dataFieldPortalData !== null && this._dataFieldPortalData.dataField instanceof MultichoiceField) {
+        //     this._dataFieldPortalData.dataField.choices = [...this._dataFieldPortalData.dataField.choices, {key: caze.stringId, value: ''}];
+        // }
     }
 }