+++ CHANGELOG.md
@@ -9,10 +9,17 @@ Full Changelog: [https://github.com/netgrif/components/commits/v6.4.0](https://g
 
 ## [6.4.0](https://github.com/netgrif/components/releases/tag/v6.4.0) (2023-06-26)
 
+### Fixed
+- [NAE-1915] TaskRef behaviour handling for multiple level
+- [NAE-1908] NAE-1906 Improvements
+
 ### Added
 - [NAE-1890] Data field component register
 - [NAE-1901] Taskref rendering update
 - [NAE-1904] CaseRef list component
+- [NAE-1497] Frontend Actions
+- [NAE-1497] Frontend Actions
+- [NAE-1904] Case ref as multichoice and enumeration
 
 ## [6.3.0](https://github.com/netgrif/components/releases/tag/v6.3.0) (2023-05-26)